Remove console/commands sub-dirs
