+++ pyproject.toml
@@ -33,7 +33,7 @@ sphinx-rtd-theme = "^2.0"
 setuptools = "^65.6.3"
 
 [tool.poetry.scripts]
-quantaq-py = "quantaq_py.console:main"
+quantaq-py = "quantaq_py.cli:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
